fix: move directory-meta route out of /project subrouter to avoid Hono :projectID path collision
