+++ packages/opencode/src/server/routes/project.ts
@@ -130,47 +130,6 @@ export const ProjectRoutes = lazy(() =>
         return c.json(next)
       },
     )
-    .post(
-      "/directory-meta",
-      describeRoute({
-        summary: "Update directory metadata",
-        description:
-          "Update name and icon for a plain directory (no git project entry). Persisted in project_recent table.",
-        operationId: "project.updateDirectoryMeta",
-        responses: {
-          200: {
-            description: "Updated directory metadata",
-            content: {
-              "application/json": {
-                schema: resolver(Project.RecentInfo),
-              },
-            },
-          },
-          ...errors(400),
-        },
-      }),
-      validator(
-        "json",
-        z.object({
-          directory: z.string(),
-          name: z.string().optional(),
-          icon: z
-            .object({
-              url: z.string().optional(),
-              color: z.string().optional(),
-            })
-            .optional(),
-        }),
-      ),
-      async (c) => {
-        const body = c.req.valid("json")
-        await Project.updateDirectoryMeta(body)
-        const list = Project.recentList()
-        const item = list.find((i) => i.kind === "directory" && i.directory === body.directory)
-        if (!item) return c.json(null, 404)
-        return c.json(item)
-      },
-    )
     .patch(
       "/:projectID",
       describeRoute({
